Vagrant tests (#112)
