+++ .github/workflows/redis.yml
@@ -111,7 +111,7 @@ jobs:
       arch: x86_64
       target: deb
       build_deps: apt-get update && apt-get install -y build-essential libssl-dev python3 python3-pip
-      packaging_deps: sudo apt-get install -y unzip dpkg-sig
+      packaging_deps: sudo apt-get install -y unzip dpkg-sig curl gnupg2
       redisversion: 7.0.0
     secrets:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
@@ -129,7 +129,7 @@ jobs:
       arch: x86_64
       target: deb
       build_deps: apt-get update && apt-get install -y build-essential libssl-dev python3 python3-pip
-      packaging_deps: sudo apt-get install -y unzip dpkg-sig
+      packaging_deps: sudo apt-get install -y unzip dpkg-sig curl gnupg2
       redisversion: 7.0.0
     secrets:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
@@ -147,7 +147,7 @@ jobs:
       arch: x86_64
       target: deb
       build_deps: apt-get update && apt-get install -y build-essential libssl-dev python3 python3-pip
-      packaging_deps: sudo apt-get install -y unzip dpkg-sig
+      packaging_deps: sudo apt-get install -y unzip dpkg-sig curl gnupg2
       redisversion: 7.0.0
     secrets:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
@@ -165,7 +165,7 @@ jobs:
       arch: x86_64
       target: rpm
       build_deps: yum install -y epel-release gcc make jemalloc-devel openssl-devel python3 python3-pip
-      packaging_deps: sudo apt-get install -y rpm unzip gpg gnupg2
+      packaging_deps: sudo apt-get install -y rpm unzip gpg gnupg2 curl
       redisversion: 7.0.0
     secrets:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
@@ -185,7 +185,7 @@ jobs:
       build_deps: |
         dnf install -y oracle-epel-release-el8
         dnf install -y gcc make jemalloc-devel openssl-devel tar git python3 python3-pip
-      packaging_deps: sudo apt-get install -y rpm unzip
+      packaging_deps: sudo apt-get install -y rpm unzip curl gnupg2
       redisversion: 7.0.0
     secrets:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
@@ -395,7 +395,7 @@ jobs:
     - name: persist redis to s3
       run: |
         s3cmd put -P redis-stack/redis-stack-server*.${{env.target}} \
-        s3://redismodules/redis-stack/snapshots/redis-stack-server-${{steps.get_version.outputs.VERSION}}-${{env.osname}}-${{env.osnick}}-${{env.arch}}.${{env.target}}
+        s3://redismodules/redis-stack/snapshots/redis-stack-server-${{steps.get_version.outputs.VERSION}}.${{env.osnick}}.${{env.arch}}.${{env.target}}
 
   # the m1 requires a zip file so that homebrew can unpack it
   build-package-osx-m1:
@@ -643,7 +643,7 @@ jobs:
           invoke dockergen -d ${{ matrix.package }}
       - name: build docker
         run: |
-          docker build -f dockers/Dockerfile.${{ matrix.package }} -t redis/${{ matrix.package }}:testing .
+          docker build -f envs/dockers/Dockerfile.${{ matrix.package }} -t redis/${{ matrix.package }}:testing .
 
       - name: test docker build
         run: |
