+++ envs/vagrants/bionic/Vagrantfile
@@ -1,0 +1,17 @@
+# -*- mode: ruby -*-
+#
+Vagrant.configure("2") do |config|
+  config.vm.box = "ubuntu/bionic64"
+  config.vm.synced_folder "../../..", "/data"
+  config.vm.network "forwarded_port", guest: 6379, host: 6379
+
+  config.vm.provider "virtualbox" do |vb|
+    vb.memory = "2048"
+  end
+
+  config.vm.provision "shell", inline: <<-SHELL
+    apt-get update
+    sudo apt install -y gdebi-core
+  SHELL
+
+end
