+++ envs/vagrants/centos7/Vagrantfile
@@ -1,0 +1,17 @@
+# vi: set ft=ruby :
+#
+Vagrant.configure("2") do |config|
+  config.vm.box = "bento/centos-7"
+  config.vm.synced_folder "../../..", "/data"
+  config.vm.network "forwarded_port", guest: 6379, host: 6379
+
+  config.vm.provider "virtualbox" do |vb|
+    vb.memory = "2048"
+  end
+
+  config.vm.provision "shell", inline: <<-SHELL
+    yum install -y epel-release
+    yum install -y python3 python3-devel python3-pip
+  SHELL
+
+end
