+++ etc/package/postinstall
@@ -5,8 +5,3 @@ for i in redis-cli redis-benchmark redis-check-aof redis-check-rdb redis-cli red
         ln -s /opt/redis-stack/bin/$i /usr/bin/$i
     fi
 done
-
-if [ -f /usr/bin/systemctl ]; then
-    systemctl enable redis-stack-server
-    systemctl start redis-stack-server
-fi
