+++ pyproject.toml
@@ -6,15 +6,16 @@ authors = ["Your Name <you@example.com>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = ">=3.6.2,<4.0"
 PyYAML = "^6.0"
+isort = "^5.10.1"
 
 [tool.poetry.dev-dependencies]
 invoke = "^1.6.0"
 requests = "^2.27.1"
 loguru = "^0.6.0"
 Jinja2 = "^3.0.3"
-redis = "^4.1.3"
+redis = "^4.3.1"
 pytest = "^7.0.0"
 docker = "^5.0.3"
 flake8 = "^4.0.1"
@@ -27,6 +28,7 @@ markers = [
     "focal: ubuntu focal tests",
     "rhel7: RedHat 7 (rhel7) tests",
     "rhel8: RedHat 8 (rhel8) tests",
+    "physical: run tests within a vagrant image",
     "dockers_redis_stack: test the redis_stack docker",
     "dockers_redis_stack_server: test the redis_stack_server docker",
     "macos: osx specific tests",
