+++ stack/__main__.py
@@ -1,7 +1,7 @@
-from optparse import OptionParser
 import sys
-from loguru import logger
+from optparse import OptionParser
 
+from loguru import logger
 
 if __name__ == "__main__":
     p = OptionParser()
@@ -105,7 +105,8 @@ if __name__ == "__main__":
 
     # a = Package(opts.OSNICK, opts.ARCH, opts.OSNAME)
     if opts.PACKAGE == "redis-stack-server":
-        from stack.recipes.redis_stack_server import RedisStackServer as pkgklass
+        from stack.recipes.redis_stack_server import \
+            RedisStackServer as pkgklass
     elif opts.PACKAGE == "redis-stack":
         from stack.recipes.redis_stack import RedisStack as pkgklass
     elif opts.PACKAGE == "redisinsight":
