+++ stack/recipes/__init__.py
@@ -1,10 +1,12 @@
-from ..config import Config
-from loguru import logger
 import os
 import shutil
 import subprocess
-import jinja2
 import tempfile
+
+import jinja2
+from loguru import logger
+
+from ..config import Config
 
 
 class Recipe(object):
