+++ tests/smoketest/test_docker_compose.py
@@ -2,9 +2,11 @@ import subprocess
 import os
 import pytest
 import time
-from helpers import RedisInsightTestMixin, RedisTestMixin
+from mixins import RedisInsightTestMixin, RedisTestMixin
+
 
 class DockerComposeBase(RedisTestMixin, object):
+    """Test running our dockers through docker compose"""
 
     VERSION = os.getenv("VERSION", "edge")
     COMPOSEFILE = "/tmp/docker-compose.yml"
@@ -15,13 +17,14 @@ class DockerComposeBase(RedisTestMixin, object):
         with open(cls.COMPOSEFILE, "w") as fp:
             fp.write(content)
 
-        cmd = ['docker-compose', '-f', cls.COMPOSEFILE, 'up', '-d']
+        cmd = ["docker-compose", "-f", cls.COMPOSEFILE, "up", "-d"]
         subprocess.run(cmd)
 
         # coalesce
         c = 0
         while c < 10:
             import redis
+
             r = redis.Redis()
             try:
                 assert r.ping()
@@ -30,12 +33,15 @@ class DockerComposeBase(RedisTestMixin, object):
                 time.sleep(1)
             c += 1
             if c == 9:
-                raise redis.exceptions.ConnectionError("Could not connect to redis, after waiting")
+                raise redis.exceptions.ConnectionError(
+                    "Could not connect to redis, after waiting"
+                )
 
     @classmethod
     def teardown_class(cls):
-        cmd = ['docker-compose', '-f', cls.COMPOSEFILE, 'down']
+        cmd = ["docker-compose", "-f", cls.COMPOSEFILE, "down"]
         subprocess.run(cmd)
+
 
 @pytest.mark.dockers_redis_stack_server
 class TestDockerComposeRedisStackServer(DockerComposeBase):
@@ -53,7 +59,8 @@ services:
       - 6379:6379
 """
         return content
-    
+
+
 @pytest.mark.dockers_redis_stack_server
 class TestDockerComposeRedisStackServerEnvVars(DockerComposeBase):
 
@@ -72,6 +79,7 @@ services:
         REDIS_ARGS: "--maxmemory 100mb"
 """
         return content
+
 
 @pytest.mark.dockers_redis_stack
 class TestDockerComposeRedisStackEnvVars(DockerComposeBase):
