+++ tests/smoketest/test_tars.py
@@ -1,34 +1,20 @@
 import docker
 import pytest
+from helpers import ROOT
+from env import DockerTestEnv
+from mixins import RedisPackagingMixin, RedisTestMixin
 
-from helpers import InDockerTestEnv, ROOT
 
+class TARTestBase(DockerTestEnv, RedisPackagingMixin, RedisTestMixin, object):
+    """Tests, in dockers for the downloadable tarballs"""
 
-class TARTestBase(InDockerTestEnv, object):
-    @classmethod
-    def setup_class(cls):
-        cls.env = docker.from_env()
+    def uninstall(self, container):  # no relevence here
+        pass
 
-        # cleanup attempt, in case one was running previously
-        try:
-            cls.teardown_class()
-        except Exception:
-            pass
+    def install(self, container):
 
-        m = docker.types.Mount("/build", ROOT, read_only=True, type="bind")
-        container = cls.env.containers.run(
-            image=cls.DOCKER_NAME,
-            name=cls.CONTAINER_NAME,
-            detach=True,
-            mounts=[m],
-            command="sleep 1200",
-            publish_all_ports=True,
-            ports={"6379/tcp": 6379},
-        )
-        cls.__CONTAINER__ = container
-
-        if getattr(cls, "__precommands__", None):
-            for cmd in cls.__precommands__(cls):
+        if getattr(self, "__precommands__", None):
+            for cmd in self.__precommands__(self):
                 out, run = container.exec_run(cmd)
                 assert out == 0
 
@@ -59,12 +45,6 @@ class TARTestBase(InDockerTestEnv, object):
 
         # validate the path now exists
         res, out = container.exec_run("ls /opt/redis-stack")
-        assert res == 0
-
-    # need to override this test for the tar case because system integration
-    # does not apply
-    def test_config_present(self):
-        res, out = self.container.exec_run("ls /opt/redis-stack/etc/redis-stack.conf")
         assert res == 0
